requirements.txt
fastapi>=0.110
uvicorn>=0.29
recipe-scrapers>=14.0
beautifulsoup4>=4.12
requests>=2.31
bcrypt>=4.0
PyJWT>=2.8
slowapi>=0.1
